LC-392: update counters in publisher to be longs instead of ints

+++ lucille-core/src/main/java/com/kmwllc/lucille/core/PublisherImpl.java
@@ -45,12 +45,12 @@ public class PublisherImpl implements Publisher {
 
   // the actual number of documents sent for processing, which may be smaller
   // than the number of calls to publish() if isCollapsing==true
-  private int numPublished = 0;
+  private long numPublished = 0;
 
-  private int numCreated = 0;
-  private int numFailed = 0;
-  private int numSucceeded = 0;
-  private int numDropped = 0;
+  private long numCreated = 0;
+  private long numFailed = 0;
+  private long numSucceeded = 0;
+  private long numDropped = 0;
 
   private Instant start;
   private final Timer timer;
@@ -269,32 +269,32 @@ public class PublisherImpl implements Publisher {
   }
 
   @Override
-  public int numPending() {
+  public long numPending() {
     return docIdsToTrack.size();
   }
 
   @Override
-  public int numPublished() {
+  public long numPublished() {
     return numPublished;
   }
 
   @Override
-  public int numCreated() {
+  public long numCreated() {
     return numCreated;
   }
 
   @Override
-  public int numSucceeded() {
+  public long numSucceeded() {
     return numSucceeded;
   }
 
   @Override
-  public int numFailed() {
+  public long numFailed() {
     return numFailed;
   }
 
   @Override
-  public int numDropped() {
+  public long numDropped() {
     return numDropped;
   }
 }
